refactor: Remove unused country-state-city files and update import paths
