+++ client/src/pages/Authentication/Signup/RegisterBusiness/RegisterBusiness.tsx
@@ -26,7 +26,7 @@ import OperationInfo from "./OperationInfo";
 import axios from "axios";
 import Modal from "react-modal";
 import Button from "@components/ui/button";
-import { Country, State, City } from "../../../../../country-sate-city";
+import { Country, State, City } from "../../../../helpers/country-sate-city";
 import * as yup from "yup";
 import { FILTERED_COUNTRY } from "@/utils/business-profile-utils";
 import { useBusinessCtx } from "@context/BusinessCtx";
